Semana6

+++ vaasschool.sql
@@ -1,7 +1,7 @@
 create database vaasschool;
 
 create table `Category` (
-  `id` int PRIMARY KEY AUTO_INCREMENT,
+  `id` bigint PRIMARY KEY AUTO_INCREMENT,
   `name` varchar(75) not null,
   `code` varchar(255) not null unique,
   `description` varchar(255) not null,
@@ -13,19 +13,19 @@ create table `Category` (
 );
 
 create table `Subcategory` (
-  `id` int PRIMARY KEY AUTO_INCREMENT,
+  `id` bigint PRIMARY KEY AUTO_INCREMENT,
   `name` varchar(75) not null,
   `code` varchar(255) not null unique,
   `description` varchar(255) not null,
   `explanatory_guide` varchar(255),
   `active` bit(1) default 0,
   `order_visualization` int not null,
-  `category_id` int not null,
+  `category_id` bigint not null,
   FOREIGN KEY(`category_id`) REFERENCES `Category`(`id`)
 );
 
 create table `Course`(
-  `id` int PRIMARY KEY AUTO_INCREMENT,
+  `id` bigint PRIMARY KEY AUTO_INCREMENT,
   `name` varchar(75) not null,
   `code` varchar (255) not null unique,
   `estimated_time_to_finish` smallint not null,
@@ -34,62 +34,62 @@ create table `Course`(
   `instructor_name` varchar(75) not null,
   `summary` text,
   `learned_skills` varchar(255),
-  `subcategory_id` int not null,
+  `subcategory_id` bigint not null,
   FOREIGN KEY(`subcategory_id`) REFERENCES `Subcategory`(`id`)
 );
 
 create table `Section`(
-  `id` int PRIMARY KEY AUTO_INCREMENT,
+  `id` bigint PRIMARY KEY AUTO_INCREMENT,
   `name` varchar(75) not null,
   `code` varchar(255) not null unique,
   `active` bit(1) default 0,
   `test` bit(1) default 0,
   `order_visualization` int,
-  `course_id` int not null,
+  `course_id` bigint not null,
   FOREIGN KEY(`course_id`) REFERENCES `Course`(`id`)
 );
 
 create table `Activity`(
-  `id` int PRIMARY KEY AUTO_INCREMENT,
+  `id` bigint PRIMARY KEY AUTO_INCREMENT,
   `title` varchar(255) not null,
   `code` varchar(255) not null unique,
   `order_visualization` int,
   `active` bit(1) default 0,
   `type` ENUM('Explanation', 'Video', 'Question'),
-  `section_id` int not null,
+  `section_id` bigint not null,
   FOREIGN KEY(`section_id`) REFERENCES `Section`(`id`)
 );
 
 create table `Explanation`(
-  `id` int PRIMARY KEY AUTO_INCREMENT,
+  `id` bigint PRIMARY KEY AUTO_INCREMENT,
   `text` varchar(255),
-  `activity_id` int not null,
+  `activity_id` bigint not null,
   FOREIGN KEY(`activity_id`) REFERENCES `Activity`(`id`)
 );
 
 create table `Video`(
-  `id` int PRIMARY KEY AUTO_INCREMENT,
+  `id` bigint PRIMARY KEY AUTO_INCREMENT,
   `url` varchar(255) not null,
-  `minutes` smallint,
+  `minutes` int,
   `transcription` varchar(255),
-  `activity_id` int not null,
+  `activity_id` bigint not null,
   FOREIGN KEY(`activity_id`) REFERENCES `Activity`(`id`)
 );
  
 create table `Question`(
-  `id` int PRIMARY KEY AUTO_INCREMENT,
+  `id` bigint PRIMARY KEY AUTO_INCREMENT,
   `statiment` varchar(255),
   `question_type` ENUM('SINGLE_ANSWER','MULTIPLE_ANSWERS','TRUE_FALSE') default 'SINGLE_ANSWER',
-  `activity_id` int not null,
+  `activity_id` bigint not null,
   FOREIGN KEY(`activity_id`) REFERENCES `Activity`(`id`)
 );
 
 create table `Alternative`(
-  `id` int PRIMARY KEY AUTO_INCREMENT,
+  `id` bigint PRIMARY KEY AUTO_INCREMENT,
   `text` varchar(255) not null,
   `order_visualization` int,
   `correct` bit(1),
   `justification` varchar(255),
-  `question_id` int not null,
+  `question_id` bigint not null,
   FOREIGN KEY(`question_id`) REFERENCES `Question`(`id`)
 );
